requirements.txt
# Core dependencies for PDF processing
PyMuPDF>=1.18.0
pdfplumber>=0.7.0
camelot-py[cv]>=0.10.1
pytesseract>=0.3.8
pdf2image>=1.16.0

# Data processing and validation
jsonschema>=4.0.0
numpy>=1.20.0
pillow>=8.0.0
opencv-python>=4.5.0

# Utilities
click>=7.0
tqdm>=4.45.0

# For development and testing
pytest>=6.0.0
ruff>=0.0.85
